llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "LogSage LLM analysis service: an AI-native Kubernetes sidekick that watches your pods, reads the logs, and turns failures into clear fixes—before they become outages. FastAPI + Strands."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "strands-agents>=1.0.0",
]

[project.urls]
Repository = "https://github.com/adefemi171/LogSage"

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_service*"]
